docs: sync integration audits with current codebase
