+++ client/pages/UnifiedReports.tsx
@@ -48,6 +48,7 @@ import {
 import HeatmapGrid, { HeatmapData } from "../components/charts/HeatmapGrid";
 import { useComplaintTypes } from "../hooks/useComplaintTypes";
 import { getAnalyticsData, getHeatmapData } from "../utils/reportUtils";
+import type { AnalyticsData, FilterOptions } from "../types/reports";
 // Recharts components will be loaded dynamically to prevent module loading issues
 import {
   CalendarDays,
@@ -73,57 +74,6 @@ import {
   Info,
 } from "lucide-react";
 // date-fns and export utilities will be loaded dynamically
-
-export interface AnalyticsData {
-  complaints: {
-    total: number;
-    resolved: number;
-    pending: number;
-    overdue: number;
-  };
-  sla: {
-    compliance: number;
-    avgResolutionTime: number;
-    target: number;
-  };
-  trends: Array<{
-    date: string;
-    complaints: number;
-    resolved: number;
-    slaCompliance: number;
-  }>;
-  wards: Array<{
-    id: string;
-    name: string;
-    complaints: number;
-    resolved: number;
-    avgTime: number;
-    slaScore: number;
-  }>;
-  categories: Array<{
-    name: string;
-    count: number;
-    avgTime: number;
-    color: string;
-  }>;
-  performance: {
-    userSatisfaction: number;
-    escalationRate: number;
-    firstCallResolution: number;
-    repeatComplaints: number;
-  };
-}
-
-export interface FilterOptions {
-  dateRange: {
-    from: string;
-    to: string;
-  };
-  ward: string;
-  complaintType: string;
-  status: string;
-  priority: string;
-}
 
 const UnifiedReports: React.FC = () => {
   const { user } = useAppSelector((state) => state.auth);
