+++ docs/architecture/audit-artifacts/integration-fixes-pr.md
@@ -1,194 +1,17 @@
-# 🔧 Critical Integration Fixes - Smart City CMS
+# Integration Hardening Status (Synchronized)
 
-## Summary
+**Last synced:** 2025-09-24  
+**Source of truth:** [`docs/INTEGRATION_FIXES_PR.md`](../../INTEGRATION_FIXES_PR.md)
 
-This PR contains critical fixes for authentication infinite loops and integration issues identified in the comprehensive code audit. These fixes resolve system-breaking issues and improve overall stability.
+This entry mirrors the refreshed integration summary. The earlier version asserted fixes that were never applied (e.g., missing import additions). The synchronized notes now reflect the actual state of the repository.
 
-## 🚨 Critical Issues Fixed
+## Highlights
 
-### 1. Authentication Infinite Redirect Loop ✅
+- ✅ RTK Query base API consolidated around `baseQueryWithReauth` with redundant imports removed.
+- ✅ Complaint attachments served by both `/complaints/:id/attachments` and `/uploads/complaint/:id/attachment` via the same controller.
+- ✅ Dashboard utilities centralized in `client/lib` modules and actively imported.
+- ⚠️ `/api/auth/refresh` endpoint absent; frontend mutation remains unused until implemented.
+- ⚠️ Legacy Redux thunks still back several dashboards; migration planning required.
+- ⚠️ Documentation cleanup identified unused helper components awaiting removal.
 
-- **Files Changed:** `client/store/api/baseApi.ts`
-- **Issue:** Missing imports caused undefined function calls in auth handling
-- **Fix:** Added proper imports for `logout`, `setError`, and `toast` functions
-- **Impact:** Resolves complete authentication system failure
-
-### 2. Enhanced Base Query Implementation ✅
-
-- **Files Changed:** `client/store/api/baseApi.ts`
-- **Issue:** App was using basic query instead of enhanced version with error handling
-- **Fix:** Switched to `baseQueryWithReauth` for proper 401 handling and token refresh
-- **Impact:** Improved error handling and automatic logout on expired tokens
-
-### 3. Missing Utility Functions ✅
-
-- **Files Added:** `client/lib/dateUtils.ts`, `client/lib/complaintUtils.ts`
-- **Issue:** Components referenced undefined utility functions
-- **Fix:** Created comprehensive utility libraries for date formatting and complaint status handling
-- **Impact:** Eliminates runtime errors in dashboard components
-
-### 4. File Upload Endpoint Compatibility ✅
-
-- **Files Changed:** `server/routes/complaintRoutes.js`
-- **Issue:** Frontend expects `/complaints/:id/attachments` but backend provides `/uploads/complaint/:id/attachment`
-- **Fix:** Added route alias to maintain backward compatibility
-- **Impact:** File uploads now work seamlessly from frontend
-
-### 5. Missing Test Dependencies ✅
-
-- **Dependencies Added:** `jsdom`, `@vitest/ui`, `msw`, `@testing-library/dom`
-- **Issue:** Test suite completely broken due to missing dependencies
-- **Fix:** Installed all required testing dependencies
-- **Impact:** Test suite is now functional and can be executed
-
-## 📊 Integration Health Improvement
-
-| Component          | Before              | After          | Improvement |
-| ------------------ | ------------------- | -------------- | ----------- |
-| Authentication     | 20% (Broken)        | 95% (Working)  | +75%        |
-| File Uploads       | 60% (Mismatch)      | 95% (Working)  | +35%        |
-| Dashboard Utils    | 70% (Missing funcs) | 95% (Complete) | +25%        |
-| Test Suite         | 0% (Broken deps)    | 85% (Ready)    | +85%        |
-| **Overall System** | **65%**             | **90%**        | **+25%**    |
-
-## 🧪 Testing
-
-### Unit Tests
-
-- ✅ All missing dependencies installed
-- ✅ Test setup validated
-- ⏳ Ready for execution (tests can now run)
-
-### Integration Tests
-
-- ✅ Auth flow tested manually
-- ✅ File upload compatibility verified
-- ✅ Utility functions tested
-
-### E2E Tests
-
-- ⏳ Cypress tests ready for execution
-- ✅ No blocking issues identified
-
-## 🔍 Code Changes
-
-### Added Files
-
-```
-client/lib/dateUtils.ts           - Date formatting utilities
-client/lib/complaintUtils.ts      - Complaint status and type utilities
-docs/CODE_AUDIT_REPORT.md         - Comprehensive audit findings
-docs/UNINTEGRATED_FUNCTIONS.md    - Integration status documentation
-docs/INTEGRATION_FIXES_PR.md      - This PR documentation
-```
-
-### Modified Files
-
-```
-client/store/api/baseApi.ts       - Fixed auth imports and enhanced query
-server/routes/complaintRoutes.js  - Added file upload endpoint alias
-vitest.config.ts                  - Fixed test configuration
-package.json                      - Added missing dependencies
-```
-
-## 🚀 Performance Impact
-
-### Build Performance
-
-- ✅ Build time unchanged (no breaking changes)
-- ✅ Bundle size minimal increase (utility functions only)
-- ✅ No performance regressions
-
-### Runtime Performance
-
-- ✅ Auth flow now stable and performant
-- ✅ File uploads work without redirects
-- ✅ Dashboard rendering improved (no undefined function errors)
-
-## 🔒 Security Considerations
-
-### Authentication Security
-
-- ✅ Enhanced error handling prevents token leakage
-- ✅ Proper logout on 401 responses
-- ✅ No security regressions introduced
-
-### API Security
-
-- ✅ File upload endpoint maintains existing security middleware
-- ✅ No new attack vectors introduced
-- ✅ Backward compatibility doesn't compromise security
-
-## 📋 Migration Guide
-
-### For Developers
-
-1. Pull latest changes
-2. Run `npm install` to get new dependencies
-3. Test auth flow to verify fixes
-4. Run test suite with `npm run test`
-
-### For Deployment
-
-- ✅ No database migrations required
-- ✅ No environment variable changes needed
-- ✅ Backward compatible changes only
-
-## 🔄 Future Improvements
-
-### Short Term (Next Sprint)
-
-1. Migrate remaining components to RTK Query for consistency
-2. Add comprehensive test coverage
-3. Implement missing refresh token endpoint
-
-### Long Term
-
-1. Performance optimization with code splitting
-2. Enhanced error monitoring
-3. API documentation updates
-
-## ✅ Checklist
-
-- [x] All critical authentication issues resolved
-- [x] File upload compatibility fixed
-- [x] Missing utility functions created
-- [x] Test dependencies installed
-- [x] Code audit documentation complete
-- [x] Integration status documented
-- [x] No breaking changes introduced
-- [x] Security implications reviewed
-- [x] Performance impact assessed
-
-## 🎯 Success Metrics
-
-### Before This PR
-
-- ❌ Users could not authenticate (infinite redirect)
-- ❌ File uploads failed with endpoint mismatch
-- ❌ Dashboard components crashed with undefined functions
-- ❌ Test suite could not execute
-
-### After This PR
-
-- ✅ Authentication works flawlessly
-- ✅ File uploads work seamlessly
-- ✅ All dashboard components render correctly
-- ✅ Test suite is functional and ready
-
-**Overall System Stability: 65% → 90% (+25% improvement)**
-
----
-
-## 📞 Support
-
-For any issues with these changes:
-
-1. Check the `docs/CODE_AUDIT_REPORT.md` for detailed analysis
-2. Review `docs/UNINTEGRATED_FUNCTIONS.md` for integration status
-3. Run the test suite to verify functionality
-4. Contact the development team for additional support
-
-## 🏆 Impact
-
-This PR transforms the Smart City CMS from a partially broken system with critical authentication failures into a fully functional, stable application ready for production use. The fixes address the root causes of user-blocking issues while maintaining backward compatibility and improving overall system reliability.
+Consult the linked primary document for full testing notes and next steps.
