+++ docs/audit-report-latest.md
@@ -1,0 +1,47 @@
+# Documentation Audit – September 2025
+
+**Generated:** 2025-09-24  
+**Scope:** `docs/` markdown review, integration validation, link and content refresh
+
+---
+
+## Summary
+
+- ✅ Reconciled integration documentation (`UNINTEGRATED_FUNCTIONS.md`, `CODE_AUDIT_REPORT.md`, `INTEGRATION_FIXES_PR.md`) with the current codebase.  
+- ✅ Updated architectural artifacts under `docs/architecture/audit-artifacts` to reference the synchronized reports instead of stale autogenerated copies.  
+- ✅ Clarified `UI_COMPONENT_MAP.md` to reflect the active `ErrorBoundary` implementation.  
+- ✅ Extended `docs/README.md` with an audit and integration index for quick navigation.  
+- ⚠️ Added backlog notes for unresolved items: missing `/api/auth/refresh`, legacy Redux thunks, unused helper components, and failing `npm run typecheck`.
+
+---
+
+## Fixed Issues
+
+| Area                              | Resolution |
+| --------------------------------- | ---------- |
+| Outdated integration counts       | Replaced templated audit text with verified findings that match the live code. |
+| Attachment endpoint mismatch note | Documented the Express alias that forwards `/complaints/:id/attachments` to the upload controller. |
+| Missing utility references        | Confirmed shared helpers in `client/lib` and updated docs accordingly. |
+| ErrorBoundary caution flag        | Reframed as an active component with documented behavior. |
+| Doc duplication in architecture artifacts | Converted files into synchronized snapshots pointing at the updated primary reports. |
+
+---
+
+## Outstanding Items (Require Engineering Work)
+
+1. Implement (or formally remove) the `/api/auth/refresh` endpoint to match the generated RTK Query mutation.  
+2. Migrate legacy `createAsyncThunk` slices (complaints, guest, admin) to RTK Query for a unified data layer.  
+3. Prune or document dormant helper components: `RoleSwitcher.tsx`, `OptimizedComponents.tsx`, `UXComponents.tsx`.  
+4. Address the pre-existing TypeScript errors so that `npm run typecheck` can pass as part of routine documentation audits.
+
+---
+
+## Follow-Up Recommendations
+
+- Schedule a backend task for refresh-token support or update the frontend to drop the unused mutation.
+- Plan a staged migration of thunk-based slices, prioritizing high-traffic dashboards.
+- Introduce a lightweight Markdown lint step once TypeScript errors are resolved to keep docs continuously validated.
+
+---
+
+_Last updated from the Smart City CMS documentation audit pipeline._
